// newcode/libs/modules/constructionsystem/common/src/lib/model/enums/quantity-query-editor/rib-function-tag.enum.ts
export enum RibFunctionTag {
	FunctionName = 'functionName',
	ParameterName = 'parameterName',
	ParameterValue = 'parameterValue',
	Operator = 'operator',
	AssignOperator = 'assignOperator',
	ParameterValueSeperator = 'parameterValueSeperator',
	StartParameterValueQuoto = 'startParameterValueQuoto',
	EndParameterValueQuoto = 'endParameterValueQuoto',
	Number = 'number',
	String = 'string',
	BlankSpace = 'blankSpace',
}